[Fixed] Do not linger unused pngs in the output dir
When using `--only_different`, some page PNGs are generated and only then
decided to be skipped. To not include them in the resulting PDF, these
are not included in the `files` array, but this also means they are
never deleted (regardless of whether `--keep_pngs` is passed).

This commit removes these pngs again immediately so they not linger
around. This also means that they are not kept with `--keep_pngs`, but
that should be ok - only pages that are in the PDF are kept now.